Initial board files for the GP20 hopperboard
This imports the variant and board.txt entries for the "Generic
G030C8Tx" from
https://github.com/stm32duino/Arduino_Core_STM32/pull/1091.

The variant is modified a bit to:
 - work with the current 1.9.0-based stm32 core (in particular,
   commented out remapped and alternative pin mappings, to be re-enabled
   later).
 - remove some ifdef checks for `ARDUINO_GENERIC_xxx` and
   `CUSTOM_PERIPHERAL_PINS` which are only needed for a generic variant.
 - remove some unneeded`extern "C"` blocks.
 - remove some unneeded defines
 - rename PinMap_TIM to PinMap_PWM (which is the old name for it in
   1.9.0).

The pin mapping itself is left unchanged, but might need some changes
later.

The boards.txt entry is a bit of a mix of the generic board and the
airid mainboard definition.