GP20-hopperboard: Stm32CubeProg SWD upload method reset board with `-hardRst`
When uploading firmware through Arduino IDE with SWD selected using
Stm32CubeProg, the board seems to stay in some sort of halt state and never runs
to user space in flash until a reset is performed. After testing it seems
that the `-g` parameter doesn't seem to work on Windows for whatever reason, but
`-hardRst` seems to work for both OS's.