GP20-Hopperboard: Reassign PB0 to TIM3
It seems that TIM1 does not work with the ArduinoCore as we use it now
(Because PCLK2 is not supported by HardwareTimer::getTimerClkFreq on G0
somehow). This also switches from the complementary output TIM1_CH2N to
the regular output TIM3_CH3, which was also not assigned to other pins
either.